convert enums to enum classes, part 3
C++11 cleanup (clean up easy enums)

This is for runtime/base/... and ended up touching a lot of files
because it turns out we have a lot of reasonably behaved enums.

+++ hphp/runtime/base/server/replay_transport.cpp
@@ -36,7 +36,7 @@ void ReplayTransport::recordInput(Transport* transport, const char *filename) {
   snprintf(buf, sizeof(buf), "%u", Process::GetThreadPid());
   hdf["tpid"] = string(buf);
 
-  hdf["cmd"] = transport->getMethod();
+  hdf["cmd"] = static_cast<int>(transport->getMethod());
   hdf["url"] = transport->getUrl();
   hdf["remote_host"] = transport->getRemoteHost();
   hdf["remote_port"] = transport->getRemotePort();
